requirements.txt
python-binance>=1.0.19
pandas>=1.5.0
numpy>=1.23.0
ta>=0.10.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
plotly>=5.10.0
python-dotenv>=0.20.0
colorlog>=6.7.0
tabulate>=0.9.0
joblib>=1.2.0
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
coverage>=7.0.0
python-telegram-bot>=20.0